requirements.txt
Django==1.9.6
django-bootstrap3==7.0.1
django-mptt==0.8.4
